refactor: split partial-upload component types into per-component types.ts
Follows the WKC per-component file layout for the two partial-upload components:

- pending-scenes-manager becomes a directory (component.ts / types.ts /
  index.ts) instead of a single file. PendingScene, UpsertPendingScene and
  IPendingScenesManager move into its types.ts.
- partial-deployments gains a types.ts holding StageDeploymentInput,
  StageDeploymentResult and IPartialDeploymentsComponent.

The six types are removed from the central src/types.ts, which now imports the
two component interfaces (type-only, so no runtime import cycle through their
component.ts factories) for AppComponents. Behavior is unchanged.
