+++ src/adapters/pending-scenes-manager/index.ts
@@ -1,0 +1,2 @@
+export { createPendingScenesManager } from './component'
+export * from './types'
