+++ src/logic/partial-deployments/index.ts
@@ -1,1 +1,2 @@
 export { createPartialDeploymentsComponent } from './component'
+export * from './types'
